edit trigger deletion, don't consider unused trigger types
