+++ Swiftfin/Views/SettingsView/UserDashboardView/EditScheduledTaskView/Components/Sections/TimeSection.swift
@@ -12,6 +12,7 @@ import SwiftUI
 extension AddTaskTriggerView {
 
     struct TimeSection: View {
+
         @Binding
         var taskTriggerInfo: TaskTriggerInfo
